assets/lecture-14/examples/flask-book/requirements.txt
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0

Flask-DebugToolbar

markdown

faker
mdgen

flake8
autopep8
djLint
python-dotenv
